+++ app/src/main/java/eu/darken/sdmse/common/areas/modules/priv/DataAppAsecModule.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.areas.modules.privdata
+package eu.darken.sdmse.common.areas.modules.priv
 
 import dagger.Binds
 import dagger.Module
@@ -20,7 +20,7 @@ import eu.darken.sdmse.common.user.UserManager2
 import javax.inject.Inject
 
 @Reusable
-class DataSystemModule @Inject constructor(
+class DataAppAsecModule @Inject constructor(
     private val userManager2: UserManager2,
     private val gatewaySwitch: GatewaySwitch,
 ) : DataAreaModule {
@@ -37,8 +37,8 @@ class DataSystemModule @Inject constructor(
             .filter { it.type == DataArea.Type.DATA && it.hasFlags(DataArea.Flag.PRIMARY) }
             .map { parentArea ->
                 DataArea(
-                    type = DataArea.Type.DATA_SYSTEM,
-                    path = parentArea.path.child("system"),
+                    type = DataArea.Type.APP_ASEC,
+                    path = parentArea.path.child("app-asec"),
                     userHandle = userManager2.systemUser().handle,
                     flags = parentArea.flags,
                 )
@@ -52,10 +52,10 @@ class DataSystemModule @Inject constructor(
 
     @Module @InstallIn(SingletonComponent::class)
     abstract class DIM {
-        @Binds @IntoSet abstract fun mod(mod: DataSystemModule): DataAreaModule
+        @Binds @IntoSet abstract fun mod(mod: DataAppAsecModule): DataAreaModule
     }
 
     companion object {
-        val TAG: String = logTag("DataArea", "Module", "DataSystem")
+        val TAG: String = logTag("DataArea", "Module", "DataAppAsec")
     }
 }
