+++ app/src/main/java/eu/darken/sdmse/common/areas/modules/priv/DataAppLibModule.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.areas.modules.privdata
+package eu.darken.sdmse.common.areas.modules.priv
 
 import dagger.Binds
 import dagger.Module
