+++ app/src/main/java/eu/darken/sdmse/common/areas/modules/priv/DataAppPrivateModule.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.areas.modules.privdata
+package eu.darken.sdmse.common.areas.modules.priv
 
 import dagger.Binds
 import dagger.Module
@@ -19,9 +19,8 @@ import eu.darken.sdmse.common.files.local.LocalGateway
 import eu.darken.sdmse.common.user.UserManager2
 import javax.inject.Inject
 
-
 @Reusable
-class DataSystemDEModule @Inject constructor(
+class DataAppPrivateModule @Inject constructor(
     private val userManager2: UserManager2,
     private val gatewaySwitch: GatewaySwitch,
 ) : DataAreaModule {
@@ -37,16 +36,13 @@ class DataSystemDEModule @Inject constructor(
         return firstPass
             .filter { it.type == DataArea.Type.DATA && it.hasFlags(DataArea.Flag.PRIMARY) }
             .map { parentArea ->
-                userManager2.allUsers().map { profile ->
-                    DataArea(
-                        type = DataArea.Type.DATA_SYSTEM_DE,
-                        path = parentArea.path.child("system_de", profile.handle.handleId.toString()),
-                        userHandle = profile.handle,
-                        flags = parentArea.flags,
-                    )
-                }
+                DataArea(
+                    type = DataArea.Type.APP_APP_PRIVATE,
+                    path = parentArea.path.child("app-private"),
+                    userHandle = userManager2.systemUser().handle,
+                    flags = parentArea.flags,
+                )
             }
-            .flatten()
             .filter {
                 val canRead = it.path.canRead(gatewaySwitch)
                 if (!canRead) log(TAG) { "Can't read $it" }
@@ -56,10 +52,10 @@ class DataSystemDEModule @Inject constructor(
 
     @Module @InstallIn(SingletonComponent::class)
     abstract class DIM {
-        @Binds @IntoSet abstract fun mod(mod: DataSystemDEModule): DataAreaModule
+        @Binds @IntoSet abstract fun mod(mod: DataAppPrivateModule): DataAreaModule
     }
 
     companion object {
-        val TAG: String = logTag("DataArea", "Module", "DataSystemDE")
+        val TAG: String = logTag("DataArea", "Module", "DataAppPrivate")
     }
 }
