+++ app/src/main/java/eu/darken/sdmse/common/areas/modules/priv/PrivateDataModule.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.areas.modules.privdata
+package eu.darken.sdmse.common.areas.modules.priv
 
 import android.os.Build
 import dagger.Binds
