+++ app/src/main/java/eu/darken/sdmse/common/areas/modules/pub/PublicObbModule.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.areas.modules.pubdata
+package eu.darken.sdmse.common.areas.modules.pub
 
 import dagger.Binds
 import dagger.Module
@@ -23,7 +23,7 @@ import eu.darken.sdmse.common.storage.SAFMapper
 import javax.inject.Inject
 
 @Reusable
-class PublicDataModule @Inject constructor(
+class PublicObbModule @Inject constructor(
     private val gatewaySwitch: GatewaySwitch,
     private val safMapper: SAFMapper,
 ) : DataAreaModule {
@@ -39,7 +39,6 @@ class PublicDataModule @Inject constructor(
                     hasApiLevel(33) -> {
                         when {
                             localGateway.hasRoot() -> {
-                                // If we have root, we need to convert any SAFPath back
                                 when (val target = parentArea.path) {
                                     is LocalPath -> target
                                     is SAFPath -> safMapper.toLocalPath(target)
@@ -53,24 +52,16 @@ class PublicDataModule @Inject constructor(
                         }
                     }
                     hasApiLevel(30) -> {
-                        val target = parentArea.path
-                        // On API30 we can do the direct SAF grant workaround
-                        when {
-                            localGateway.hasRoot() -> when (target) {
-                                is SAFPath -> safMapper.toLocalPath(target)
-                                else -> target
-                            }
-                            else -> when (target) {
-                                is LocalPath -> safMapper.toSAFPath(target)
-                                is SAFPath -> target
-                                else -> null
-                            }
+                        when (val target = parentArea.path) {
+                            is LocalPath -> safMapper.toSAFPath(target)
+                            is SAFPath -> target
+                            else -> null
                         }
                     }
                     else -> parentArea.path
                 }
-                val childPath = accessPath?.child("Android", "data") ?: return@mapNotNull null
-                parentArea to childPath
+                val dataAccessPath = accessPath?.child("Android", "obb") ?: return@mapNotNull null
+                parentArea to dataAccessPath
             }
             .filter {
                 val canRead = it.second.canRead(gatewaySwitch)
@@ -79,7 +70,7 @@ class PublicDataModule @Inject constructor(
             }
             .map { (parentArea, path) ->
                 DataArea(
-                    type = DataArea.Type.PUBLIC_DATA,
+                    type = DataArea.Type.PUBLIC_OBB,
                     path = path,
                     flags = parentArea.flags,
                     userHandle = parentArea.userHandle,
@@ -94,10 +85,10 @@ class PublicDataModule @Inject constructor(
     @InstallIn(SingletonComponent::class)
     @Module
     abstract class DIM {
-        @Binds @IntoSet abstract fun mod(mod: PublicDataModule): DataAreaModule
+        @Binds @IntoSet abstract fun mod(mod: PublicObbModule): DataAreaModule
     }
 
     companion object {
-        val TAG: String = logTag("DataArea", "Module", "Public", "Data")
+        val TAG: String = logTag("DataArea", "Module", "Public", "Obb")
     }
 }
